requirement.txt
anyio==4.12.1
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.3
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.4.1
onnxruntime==1.23.2
packaging==25.0
protobuf==6.33.4
pycparser==2.23
pynput==1.8.1
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.3
soundfile==0.13.1
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
